Add notification validation
