+++ website/docs/r/detector.html.markdown
@@ -49,7 +49,7 @@ variable "clusters" {
 As SignalFx supports different notification mechanisms a comma-delimited string is used to provide inputs. If you'd like to specify multiple notifications, then each should be a member in the list, like so:
 
 ```
-notifications = ["Email,foo-alerts@example.com", "Slack,credentialId,#channel"]
+notifications = ["Email,foo-alerts@example.com", "Slack,credentialId,channel"]
 ```
 
 This will likely be changed in a future iteration of the provider. See [SignalFX Docs](https://developers.signalfx.com/detectors_reference.html#operation/Create%20Single%20Detector) for more information. For now, here are some example of how to configure each notification type:
@@ -76,10 +76,10 @@ notifications = ["PagerDuty,credentialId"]
 
 ### Slack
 
-Include the `#` on the channel name!
+Exclude the `#` on the channel name!
 
 ```
-notifications = ["Slack,credentialId,#channel"]
+notifications = ["Slack,credentialId,channel"]
 ```
 
 ### Team
